TestTask/requirements.txt
Django==3.2.13
psycopg2-binary==2.9.3
sqlparse==0.4.2
python-dotenv~=0.20.0
google-api-python-client==2.51.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
django-gsheets==0.0.10
django-sslserver==0.22
django-sitemessage==1.3.4
celery==5.2.7
redis==4.3.3